:construction: Started work on the normalizer object
Writing to standard out from OpenCSV turns out to be a little more involved than I first thought. I got that sorted and then tucked it away inside of the Normalizer object.

I also got testing working on the Normalizer object. I did have to make a small concession on the column names of the CSV. There's probably someway to capitalize them as they were input but it didn't seem as important as making progress on the other items.